requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
SQLAlchemy
